+++ tests/use_cases/qa/test_multiple_chunk_retriever_qa.py
@@ -2,6 +2,7 @@ from pytest import fixture
 
 from intelligence_layer.connectors import QdrantInMemoryRetriever
 from intelligence_layer.core import NoOpTracer
+from intelligence_layer.core.tracer.in_memory_tracer import InMemoryTracer
 from intelligence_layer.use_cases import MultipleChunkRetrieverQa, RetrieverBasedQaInput
 
 
@@ -18,7 +19,8 @@ def test_multiple_chunk_retriever_qa_using_in_memory_retriever(
 ) -> None:
     question = "When was Robert Moses born?"
     input = RetrieverBasedQaInput(question=question)
-    output = multiple_chunk_retriever_qa.run(input, no_op_tracer)
+    tracer = InMemoryTracer()
+    output = multiple_chunk_retriever_qa.run(input, tracer)
     assert output.answer
     assert "1888" in output.answer
     assert len(output.sources) == 5