+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - feature: `run_dataset` now takes `max_workers` as an additional argument. Defaults to 10, as it was hardcoded before.
 
 ### New Features
+- feature: `Llama2InstructModel` to support llama-2 models in Aleph Alpha API
+- feature: `Llama3InstructModel` to support llama-3 models in Aleph Alpha API
+- feature: `ExpandChunks`-task caches chunked documents by ID
 
 ### Fixes